+++ doc/changelog.md
@@ -1,3 +1,9 @@
+**Beta 10:**
+* Added: setting to change the color of the frontier text.
+ 
+**Beta 9:**
+* Fixed: bug when reading chunk frontiers from file.
+ 
 **Beta 8:**
 * Fixed: error when opening frontiers from previous versions that are missing certain visibility parameters.